requirements.txt
requests
pydantic
pyyaml
databricks-sdk
snowflake
snowflake-connector-python
snowflake-snowpark-python